+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.0.0-alpha.10
 
+* Added: Anonymous getters/setters
 * Added: "Trickle-Down" function lookup: function calls will now descope until the requested variable provides a function or other callable value
 * Changed: Updated `quickscope` to 0.1.5
 * Removed: `[get]`